Ensure inputs are integer dtype for embedding layer

+++ VishwamAI/scripts/model_architecture.py
@@ -68,8 +68,8 @@ class VishwamAIModel(hk.Module):
             raise ValueError("Input must be of type `str`, `List[str]`, or `List[List[int]]`")
         print(f"Data type of inputs after conversion: {inputs.dtype}")
 
-        # Convert inputs to float32 for model parameter initialization
-        inputs = jax.numpy.array(inputs, dtype=jnp.float32)
+        # Ensure inputs are integer dtype for embedding layer and transformer initialization
+        inputs = jax.numpy.array(inputs, dtype=jnp.int32)
 
         # Initialize the parameters for the transformer
         rng = jax.random.PRNGKey(42)
